+++ R/RcppExports.R
@@ -5,11 +5,11 @@ compute_df <- function(alpha, lambda, xA) {
     .Call(`_sentometrics_compute_df`, alpha, lambda, xA)
 }
 
-compute_sentiment_bigrams <- function(texts, lexicons, how) {
-    .Call(`_sentometrics_compute_sentiment_bigrams`, texts, lexicons, how)
-}
-
 compute_sentiment_onegrams <- function(texts, lexicons, how) {
     .Call(`_sentometrics_compute_sentiment_onegrams`, texts, lexicons, how)
+}
+
+compute_sentiment_valence <- function(texts, lexicons, how) {
+    .Call(`_sentometrics_compute_sentiment_valence`, texts, lexicons, how)
 }