+++ R/sentometrics.R
@@ -7,15 +7,15 @@
 #' in standard time series analysis and text mining packages. The package also provides an interface to model
 #' model the impact of sentiment in texts on a given variable, by first computing a wide range of textual
 #' sentiment time series and then selecting those that are most informative. Altogether, \pkg{sentometrics} integrates the
-#' \emph{qualification} of sentiment from texts, the \emph{aggregation} into different sentiment measures and the optimized
+#' fast \emph{qualification} of sentiment from texts, the \emph{aggregation} into different sentiment measures and the optimized
 #' \emph{prediction} based on these measures.
 #'
 #' @section Main functions:
 #' \itemize{
-#' \item Feature generation: \code{\link{sento_corpus}}, \code{\link{add_features}}
+#' \item Corpus features generation: \code{\link{sento_corpus}}, \code{\link{add_features}}
 #' \item Sentiment computation and aggregation into sentiment measures: \code{\link{ctr_agg}},
-#' \code{\link{compute_sentiment}}, \code{\link{sento_measures}}, \code{\link{peakdocs}}, \code{\link{measures_merge}},
-#' \code{\link{measures_global}}, and a series of other \code{measures_xyz}, generic and extractor functions
+#' \code{\link{compute_sentiment}}, \code{\link{sento_measures}}, \code{\link{peakdocs}}, and a series of
+#' \code{measures_xyz}, generic and extractor functions
 #' \item Sparse modelling: \code{\link{ctr_model}}, \code{\link{sento_model}}
 #' \item Prediction and post-modelling analysis: \code{\link{predict.sentomodel}}, \code{\link{retrieve_attributions}},
 #' \code{\link{plot_attributions}}, \code{\link{perform_MCS}}
@@ -25,7 +25,7 @@
 #' The development version of the package resides at \url{https://github.com/sborms/sentometrics}.
 #'
 #' @note The methodology behind the sentiment aggregation framework can be consulted in the working paper ``Questioning
-#' the news about economic growth: Sparse forecasting using thousands of news-based sentiment values'' (Ardia, Bluteau, and
+#' the news about economic growth: Sparse forecasting using thousands of news-based sentiment values'' (Ardia, Bluteau and
 #' Boudt, 2017) at \url{https://ssrn.com/abstract=2976084}. The vignette ``The R package sentometrics
 #' to compute, aggregate and predict with textual sentiment'' (Ardia, Bluteau, Borms and Boudt, 2017) at
 #' \url{https://ssrn.com/abstract=3067734} provides a hands-on introduction to the methodology and
@@ -86,12 +86,14 @@
 #' @docType data
 #'
 #' @description
-#' A \code{list} containing all built-in valence word lists, as \code{data.table}s with two columns: a \code{x} column with the
-#' words, and a \code{y} column with the values associated to each word. The \code{list} element names indicate the language
-#' (based on the two-letter ISO code convention as in \code{\link[stopwords]{stopwords}}) of the valence word list. All
-#' non-English word lists are translated via Microsoft Translator through Microsoft Word. Only the entries whose translation
-#' differs from the original entry are kept. The valence word lists are in the form required for further sentiment analysis.
-#' All words are unigrams and in lowercase. The built-in valence word lists are available in following languages:
+#' A \code{list} containing all built-in valence word lists, as \code{data.table}s with three columns: a \code{x} column with
+#' the words, a \code{y} column with the values associated to each word, and a \code{t} column with the type of valence
+#' shifter (\code{1} = negators, \code{2} = amplifiers, \code{3} = deamplifiers). The \code{list} element names indicate the
+#' language (based on the two-letter ISO code convention as in \code{\link[stopwords]{stopwords}}) of the valence word list.
+#' All non-English word lists are translated via Microsoft Translator through Microsoft Word. Only the entries whose
+#' translation differs from the original entry are kept. The valence word lists are in the form required for further
+#' sentiment analysis. All words are unigrams and in lowercase. The built-in valence word lists are available in following
+#' languages:
 #'
 #' \itemize{
 #'   \item English (\code{"en"})