+++ tests/testthat/test_attribution.R
@@ -14,7 +14,7 @@ corpus <- quanteda::corpus_sample(
 )
 
 data("list_lexicons")
-lex <- list_lexicons[c("GI_en", "LM_en")]
+lex <- setup_lexicons(list_lexicons[c("GI_en", "LM_en")])
 ctrA <- ctr_agg(howWithin = "counts", howDocs = "proportional", howTime = "almon", by = "month",
                 lag = 7, ordersAlm = 1:3, do.inverseAlm = TRUE, do.ignoreZeros = FALSE, fill = "latest")