+++ tests/testthat/test_model_specification.R
@@ -11,7 +11,7 @@ data("usnews")
 corpus <- quanteda::corpus_sample(quanteda::corpus_subset(sento_corpus(corpusdf = usnews)), 750)
 
 data("list_lexicons")
-lex <- list_lexicons[c("GI_en", "LM_en")]
+lex <- setup_lexicons(list_lexicons[c("GI_en", "LM_en")])
 ctrA <- ctr_agg(howWithin = "proportional", howDocs = "proportional", howTime = "almon", by = "month",
                 lag = 7, ordersAlm = 1:3, do.inverseAlm = TRUE, do.ignoreZeros = FALSE, fill = "latest")
 
@@ -37,7 +37,7 @@ out1 <- sento_model(sentomeasures, y, x = x, ctr = ctrM1)
 ctrM2 <- ctr_model(model = "gaussian", type = "AIC", h = -1, alphas = c(0.2, 0.7))
 out2 <- sento_model(sentomeasures, y, x = x, ctr = ctrM2)
 
-ctrM3 <- ctr_model(model = "gaussian", type = "BIC", h = 0, alphas = c(0.2, 0.7))
+ctrM3 <- ctr_model(model = "gaussian", type = "BIC", h = 0, do.regularize.x = TRUE, alphas = c(0.2, 0.7))
 out3 <- sento_model(sentomeasures, y, x = x, ctr = ctrM3)
 
 ctrM4 <- ctr_model(model = "gaussian", type = "cv", h = 3, trainWindow = nrow(x) - 30, testWindow = 7, alphas = c(0.2, 0.7))
@@ -72,6 +72,7 @@ out11 <- sento_model(measures_select(sentomeasures, c("LM_en", "wsj", "almon1"))
 test_that("Different model specifications give specified output", {
   expect_equal(N - 8, nrow(out1$x))
   expect_equal(N - 1, nrow(out2$x))
+  expect_equal(N, nrow(out3$x))
   expect_equal(N - 3, nrow(out4$x))
   expect_equal(N - 1, nrow(out5$x))
   expect_equal(N - 5, nrow(out6$x))
@@ -81,7 +82,7 @@ test_that("Different model specifications give specified output", {
   expect_true(all(out8$alphas %in% c(0, 0.4, 1)))
   expect_equal(out7$lambda, (50:1)[which(out7$ic$matrix == min(out7$ic$matrix, na.rm = TRUE), arr.ind = TRUE)[1, 2]])
   expect_equal(N - 1 - nSample - 2 + 1, length(out8$models))
-  expect_true(all(sapply(c(list(out1, out2, out3, out4, out5, out7), out8$models),
+  expect_true(all(sapply(c(list(out1, out2, out4, out5, out7), out8$models),
                          function(out) stats::coef(out$reg)[c("x1", "x2"), ]) != 0))
   expect_error(sento_model(sentomeasures, y, x = x, ctr = ctrM9))
   expect_null(summary(out1))